Revert change to add checksums to rpm purls
Checksums will make it harder to make purls of components shared between
different SBOMs match because of various differences on how the
checksums may be computed. Instead, we should just rely on the "redhat"
namespace to differentiate our packages from those that may happen to
share the same name, version, and repo ID but are distributed by a
different vendor.